requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
sqlalchemy>=2.0.0
sqlglot>=18.0.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0

# API clients for commercial models
openai>=1.0.0
google-generativeai>=0.3.0

# Utility packages
tiktoken>=0.5.0
func-timeout>=4.3.5
tqdm>=4.65.0
psutil>=5.9.0

# Optional: For better performance
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0

# Data processing
openpyxl>=3.1.0
